requirements.txt
django
djangorestframework
requests
pyTelegramBotAPI
openai
celery
django-celery-beat